:recycle: refactor : make abstract class UserRoleChecker<T>
