+++ com.rush.logistic.client.order-delivery/src/main/java/com/rush/logistic/client/order_delivery/domain/order/controller/OrderController.java
@@ -11,7 +11,7 @@ import com.rush.logistic.client.order_delivery.domain.order.service.OrderService
 import com.rush.logistic.client.order_delivery.global.auth.UserInfo;
 import com.rush.logistic.client.order_delivery.global.auth.UserInfoHeader;
 import com.rush.logistic.client.order_delivery.global.auth.UserRole;
-import com.rush.logistic.client.order_delivery.global.auth.checker.UserRoleChecker;
+import com.rush.logistic.client.order_delivery.global.auth.checker.OrderUserRoleChecker;
 import com.rush.logistic.client.order_delivery.global.response.BaseResponse;
 import com.rush.logistic.client.order_delivery.domain.order.exception.OrderCode;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.UUID;
 
 @Slf4j
@@ -30,12 +29,12 @@ import java.util.UUID;
 public class OrderController {
     private final OrderCreateService orderCreateService;
     private final OrderService orderService;
-    private final UserRoleChecker userRoleChecker;
+    private final OrderUserRoleChecker orderUserRoleChecker;
 
     @PostMapping
     public ResponseEntity<Object> createOrder(@UserInfoHeader UserInfo userInfo, @RequestBody OrderAndDeliveryCreateReq requestDto) {
         log.info("OrderController createOrder");
-        GetUserInfoRes getUserInfoRes = userRoleChecker.getUserAndCheckAllRole(userInfo);
+        GetUserInfoRes getUserInfoRes = orderUserRoleChecker.getUserAndCheckAllRole(userInfo);
 
         OrderAllRes responseDto = orderCreateService.createDeliveryAndOrder(userInfo.getUserId(), requestDto);
         return ResponseEntity.ok().body(BaseResponse.toResponse(OrderCode.CREATE_ORDER_OK, responseDto));
@@ -44,7 +43,7 @@ public class OrderController {
     @GetMapping("/{orderId}")
     public ResponseEntity<Object> getOrderById(@UserInfoHeader UserInfo userInfo, @PathVariable UUID orderId) {
         log.info("OrderController getOrderById");
-        GetUserInfoRes getUserInfoRes = userRoleChecker.getUserAndCheckAllRole(userInfo);
+        GetUserInfoRes getUserInfoRes = orderUserRoleChecker.getUserAndCheckAllRole(userInfo);
 
         OrderAllRes responseDto = orderService.getOrderDetail(orderId, getUserInfoRes);
         return ResponseEntity.ok().body(BaseResponse.toResponse(OrderCode.GET_ORDER_OK, responseDto));
@@ -53,7 +52,7 @@ public class OrderController {
     @PatchMapping("/{orderId}")
     public ResponseEntity<Object> updateOrder(@UserInfoHeader UserInfo userInfo, @PathVariable UUID orderId, @RequestBody OrderAllReq requestDto) {
         log.info("OrderController updateOrder");
-        GetUserInfoRes getUserInfoRes = userRoleChecker.getUserAndCheckRole(Arrays.asList(UserRole.MASTER, UserRole.HUB), userInfo);
+        GetUserInfoRes getUserInfoRes = orderUserRoleChecker.getUserAndCheckRole(Arrays.asList(UserRole.MASTER, UserRole.HUB), userInfo);
 
         OrderUpdateRes responseDto = orderService.updateOrder(orderId, requestDto, getUserInfoRes);
         return ResponseEntity.ok().body(BaseResponse.toResponse(OrderCode.UPDATE_ORDER_OK, responseDto));
@@ -62,7 +61,7 @@ public class OrderController {
     @DeleteMapping("/{orderId}")
     public ResponseEntity<Object> deleteOrder(@UserInfoHeader UserInfo userInfo, @PathVariable UUID orderId) {
         log.info("OrderController deleteOrder");
-        GetUserInfoRes getUserInfoRes = userRoleChecker.getUserAndCheckRole(Arrays.asList(UserRole.MASTER, UserRole.HUB), userInfo);
+        GetUserInfoRes getUserInfoRes = orderUserRoleChecker.getUserAndCheckRole(Arrays.asList(UserRole.MASTER, UserRole.HUB), userInfo);
 
         UUID deletedId = orderService.deleteOrder(orderId, userInfo.getUserId(), getUserInfoRes);
         return ResponseEntity.ok().body(BaseResponse.toResponse(OrderCode.DELETE_ORDER_OK, OrderIdRes.toDto(deletedId)));
