+++ com.rush.logistic.client.order-delivery/src/main/java/com/rush/logistic/client/order_delivery/domain/order/service/OrderService.java
@@ -14,8 +14,7 @@ import com.rush.logistic.client.order_delivery.domain.order.domain.Order;
 import com.rush.logistic.client.order_delivery.domain.order.exception.OrderCode;
 import com.rush.logistic.client.order_delivery.domain.order.exception.OrderException;
 import com.rush.logistic.client.order_delivery.domain.order.repository.OrderRepository;
-import com.rush.logistic.client.order_delivery.global.auth.checker.UserRoleChecker;
-import jakarta.persistence.EntityManager;
+import com.rush.logistic.client.order_delivery.global.auth.checker.OrderUserRoleChecker;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -32,11 +31,11 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final DeliveryRepository deliveryRepository;
     private final DeliveryRouteRepository deliveryRouteRepository;
-    private final UserRoleChecker userRoleChecker;
+    private final OrderUserRoleChecker orderUserRoleChecker;
 
     public OrderAllRes getOrderDetail(UUID orderId, GetUserInfoRes getUserInfoRes) {
         Order order = getOrderEntityById(orderId);
-        userRoleChecker.checkInCharge(order, getUserInfoRes);
+        orderUserRoleChecker.checkInCharge(order, getUserInfoRes);
 
         List<DeliveryRoute> deliveryRoutes = deliveryRouteRepository.findAllByDelivery(order.getDelivery());
         return OrderAllRes.fromEntity(order, deliveryRoutes);
@@ -48,7 +47,7 @@ public class OrderService {
                 .orElseThrow(() -> new DeliveryException(DeliveryCode.DELIVERY_NOT_EXIST));
 
         Order order = getOrderEntityById(orderId);
-        userRoleChecker.checkInCharge(order, getUserInfoRes);
+        orderUserRoleChecker.checkInCharge(order, getUserInfoRes);
 
         order.updateAll(requestDto, delivery);
         orderRepository.saveAndFlush(order);
@@ -60,7 +59,7 @@ public class OrderService {
     @Transactional
     public UUID deleteOrder(UUID orderId, Long userId, GetUserInfoRes getUserInfoRes) {
         Order order = getOrderEntityById(orderId);
-        userRoleChecker.checkInCharge(order, getUserInfoRes);
+        orderUserRoleChecker.checkInCharge(order, getUserInfoRes);
 
         order.softDelete(userId.toString());
         return order.getId();
